ai_service/requirements.txt
Django==4.2
djangorestframework==3.14.0
torch==2.1.0
torchvision==0.16.0
timm==0.9.12
pika==1.3.2
Pillow==10.1.0
numpy==1.24.3
requests==2.31.0
